# openapi.yaml
openapi: 3.0.3
info:
  title: Kaspa TN12 Dashboard API
  version: 1.0.0
  description: Backend API for Kaspa Testnet 12 Dashboard with Atomic Swaps and Deadman Switch
paths:
...
  /api/atomic-swap/initiate:
    post:
      summary: Initiate an atomic swap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [direction, amount]
              properties:
                direction: { type: string, enum: [kas2eth, eth2kas] }
                kasAddress: { type: string }
                amount: { type: number }
                timelock: { type: number }
                rate: { type: number }
                counterparty: { type: string }
      responses:
        '200':
          description: Swap initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  swapId: { type: string }
                  preimage: { type: string }
                  hashlock: { type: string }
                  htlcAddress: { type: string }
                  timelock: { type: number }
                  status: { type: string }
  /api/atomic-swap/claim:
    post:
      summary: Claim an atomic swap with preimage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [swapId, preimage]
              properties:
                swapId: { type: string }
                preimage: { type: string }
      responses:
        '200':
          description: Swap claimed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
  /api/atomic-swap/refund:
    post:
      summary: Refund an atomic swap after timelock expiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [swapId]
              properties:
                swapId: { type: string }
      responses:
        '200':
          description: Swap refunded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
...